[test-hypertable] Improve unit tests for local hypertables, including unit tests of specific behaviors.  Move unit tests basd on old hash function to separate file.  Clean up unit-testing code overall.
